pyproject.toml
[project]
name = "singularity-parser"
version = "0.1.0"
description = ""
authors = [
    {name = "comavp",email = "comavp@yandex.ru"}
]
requires-python = ">=3.10,<4.0"
dependencies = [
    "loguru (>=0.7.3,<0.8.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
